Live audit of the full 31,117-row EU feed: saleRate spans 0-1, soldPerDay
spans 0-965,709.12 (both decimal, not integers), and 53 item names contain
commas inside quoted fields with RFC 4180 escapes.

Naive comma splitting silently shifted every column after a quoted name, so
those items reported another column's value as their price -- plausible and
wrong, with no error. Now parsed with the csv crate. soldPerDay is stored as
REAL (migration 0032).

Fixtures are real rows and pin marketValue=470449678 for item 94582, which
only holds if quoting is handled correctly.

+++ crates/storage/src/sqlite/tsm.rs
@@ -44,7 +44,7 @@ impl TsmRepository for SqliteTsm {
             .bind(sample.historical.get() as i64)
             .bind(sample.avg_sale_price.get() as i64)
             .bind(sample.sale_rate_bp as i64)
-            .bind(sample.sold_per_day as i64)
+            .bind(sample.sold_per_day)
             .bind(sample.updated_at.get() as i64)
             .execute(&mut *tx)
             .await
